Src/llm-service/requirements.txt
fastapi==0.129.1
uvicorn==0.41.0
python-dotenv==1.2.1
langchain==1.2.10
langchain-community==0.4.1
langchain-groq==1.1.2
langchain-experimental==0.4.1
langchain-huggingface==1.2.0
sentence-transformers==5.2.3
pydantic-settings>=2.0.0
pytesttenacity>=8.0.0
